scripts/requirements.txt
py-solc-x>=1.1.0
web3>=6.11.3
requests>=2.31.0
